crypto_project/requirements.txt
Flask==2.2.3
Flask-SocketIO==5.2.0
PyMySQL==1.0.3
python-dotenv==1.0.0
ipython==8.12.0
cryptography==40.0.2
